move logic to mapper

+++ features/content/play_broadcaster/src/main/java/com/tokopedia/play/broadcaster/domain/repository/PlayBroProductRepository.kt
@@ -21,6 +21,6 @@ interface PlayBroProductRepository {
     ): List<ProductUiModel>
 
     suspend fun getProductTagSummarySection(
-        channelID: Int,
+        channelID: Long,
     ): List<ProductTagSectionUiModel>
 }
